Add balance display

+++ AndroidApp/app/src/main/java/com/androbrain/androidapp/data/repository/BudgetModel.kt
@@ -1,10 +1,8 @@
 package com.androbrain.androidapp.data.repository
 
-import androidx.room.PrimaryKey
-
 data class BudgetModel(
     val id: Int,
-    val amount: Int,
+    val amount: Long,
     val description: String,
     val isSpending: Boolean,
 )
